Refactor: firebase 관련 파일 이동
shared/utils/fcm 디렉토리로 이동함
