+++ src/pages/home/Home.vue
@@ -104,12 +104,12 @@
 import trip_image from '@/assets/trip_image.png'
 import { friendFeedMock, tripRecommendationListMockData } from '@/entities/trip/trip.mock'
 import FriendActivityFeed from '@/features/trip/MyTrip/ui/FriendActivityFeed.vue'
-import { initFCM } from '@/initFCM'
 import Card from '@/shared/components/atoms/card/Card.vue'
 import TypographyCaption from '@/shared/components/atoms/typography/TypographyCaption.vue'
 import TypographyHead1 from '@/shared/components/atoms/typography/TypographyHead1.vue'
 import TypographyHead2 from '@/shared/components/atoms/typography/TypographyHead2.vue'
 import TypographySubTitle1 from '@/shared/components/atoms/typography/TypographySubTitle1.vue'
+import { initFCM } from '@/shared/utils/fcm/initFCM'
 import { computed } from 'vue'
 import { useRouter } from 'vue-router'
